Remove PaginatedResource, favor new paginators
Remove PaginatedResource and remove tests for it. Replace with initial
unit testing for HasNextPaginator, as part of the new suite of
paginators.
